% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_clr.R
\name{bayes_clr}
\alias{bayes_clr}
\title{Fits a Bayesian conditional logistic regression model}
\usage{
bayes_clr(
  x,
  y,
  strata,
  num_burn = 2500,
  num_thin = 1,
  num_save = 2500,
  update_sigma_beta = TRUE,
  sigma_beta = 1,
  seed = 1996
)
}
\arguments{
\item{x}{A matrix of covariates to include in the linear predictor component.}

\item{y}{A vector of outcomes (binary).}

\item{strata}{An integer vector indicating the strata which each observation
belongs to. Prior to calling \code{bayes_clr}, all data components must be
ordered by strata, and strata should be in increasing order.}

\item{num_burn}{Number of burn-in iterations for the chain.}

\item{num_thin}{Post burn-in thinning interval for the chain.}

\item{num_save}{Number of posterior samples to save; in total, \code{num_burn + num_save * num_thin} iterations are run.}

\item{update_sigma_beta}{If \code{TRUE}, \code{sigma_beta} is updated to encourage optimal M-H acceptance rates for the vector of fixed coefficients.}

\item{sigma_beta}{Proposal distribution variance for fixed effect coefficients.}

\item{seed}{Random seed for reproducibility.}
}
\value{
Returns a list with the following components:
\itemize{
\item \code{beta}: \code{num_save} x \code{ncol(x)} matrix of regression coefficients.
\item \code{loglik}: vector of length \code{num_save} of the total log-likelihood.
\item \code{time}: total time taken for the model to run (in minutes).
\item \code{WAIC}: model WAIC (Widely Applicable Information Criteria).
}
}
\description{
Runs a conventional conditional logistic regression model (similar to
\code{clogit} in the \code{survival} package) for data which have been
processed using the standard (time-stratified) case-crossover approach.
}
\examples{
# TBD
}
