% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clbart-package.R
\docType{package}
\name{clbart-package}
\alias{clbart-package}
\alias{_PACKAGE}
\title{clbart: Conditional Logistic Regression with BART}
\description{
Fit a Bayesian conditional logistic regression model with a single regression coefficient that is allowed to vary across the covariate space via the the Bayesian Additive Regression Trees (BART) framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jacobenglert/clbart}
  \item Report bugs at \url{https://github.com/jacobenglert/clbart/issues}
}

}
\author{
\strong{Maintainer}: Jacob Englert \email{jacobenglert@outlook.com} (\href{https://orcid.org/0000-0001-5416-3604}{ORCID})

}
\keyword{internal}
