% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clbart.R
\name{clbart}
\alias{clbart}
\title{Estimate Heterogeneous Effects with Conditional Logistic Regression}
\usage{
clbart(
  w,
  x = NULL,
  y,
  z,
  stratum,
  num_trees = 5,
  seed = 2187,
  iter = 10000,
  thin = 1,
  warmup = floor(iter/2),
  sigma2_beta = 1,
  sigma2_beta_update_freq = 100,
  beta_acc_prob = 0.35,
  n_min = 5,
  moves = c("grow", "prune", "change"),
  move_probs = c(0.3, 0.3, 0.4),
  alpha_rho = 0.95,
  beta_rho = 2,
  alpha_sigma = 1e-04,
  beta_sigma = 1e-04
)
}
\arguments{
\item{w}{A data frame of potential effect moderators.}

\item{x}{A data frame of effect confounders.}

\item{y}{A numeric vector of 1's and 0's corresponding to the outcome}

\item{z}{A numeric vector corresponding to the primary treatment or exposure
of interest. May be continuous or binary.}

\item{stratum}{A vector indicating the strata each observation belongs to.}

\item{num_trees}{integer number of trees to include in the BART portion of
the model}

\item{seed}{A random seed to initiate the algorithm with.}

\item{iter}{The total number of MCMC iterations to run.}

\item{thin}{Thinning rate for posterior samples.}

\item{warmup}{Number of MCMC iterations to run before storing posterior
samples.}

\item{sigma2_beta}{Initial value of the variance parameter used in the joint
multivariate normal proposal distribution for the confounding parameters.
This is updated during the warm-up period every \code{sigma2_beta_update_freq}
iterations.}

\item{sigma2_beta_update_freq}{Frequency with which to update \code{sigma2_beta}
during the warm-up period.}

\item{beta_acc_prob}{Target acceptance probability for the joint
Metropolis-Hastings proposal for the confounding parameters. \code{sigma2_beta}
will be updated throughout the MCMC with the goal of achieving this rate.}

\item{n_min}{Minimum number of cases for proposed terminal nodes. Moves that
would result in fewer than \code{n_min} cases in a leaf node are not considered.}

\item{moves}{Possible moves for the reversible-jump algorithm (DO NOT TOUCH).}

\item{move_probs}{Prior probabilities corresponding to the moves in \code{moves}.}

\item{alpha_rho, beta_rho}{Parameters for the tree splitting prior as defined
in Chipman et al. (2010).}

\item{alpha_sigma, beta_sigma}{Shape and scale parameters for the inverse
gamma prior on the terminal node variance.}
}
\description{
Estimate Heterogeneous Effects with Conditional Logistic Regression
}
\examples{
'INSERT EXAMPLE HERE'

}
\references{
Chipman, H. A., George, E. I., and McCulloch, R. E. (2010). BART: Bayesian additive
regression trees. The Annals of Applied Statistics, 4(1).
}
