% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_predictions.R
\name{get_forest_posterior_predictions}
\alias{get_forest_posterior_predictions}
\title{Get predictions for a set of predictors given posterior samples of forests}
\usage{
get_forest_posterior_predictions(forest_posterior, x)
}
\arguments{
\item{forest_posterior}{List of forest data frames returned by the \code{clbart} function.}

\item{x}{Matrix of covariates to make predictions for.}
}
\value{
Returns a matrix of predictions with \code{nrow(x)} rows and \code{opts$num_save} columns, where opts$num_save is the number of posterior samples taken during model-fitting.
}
\description{
Get predictions for a set of predictors given posterior samples of forests
}
\examples{
# fit is an object returned by clbart()
# w is a matrix of covariates used in the BART component
# pred_post <- get_forest_posterior_predictions(fit$forest, w)
}
