% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clbart.R
\name{Hypers}
\alias{Hypers}
\title{Create a list of hyperparameter values}
\usage{
Hypers(
  num_tree = 50,
  k = 0.5,
  base = 0.95,
  power = 2,
  mu_mu = 0,
  sigma_mu = NULL,
  s = NULL,
  alpha = 1,
  alpha_scale = NULL,
  alpha_shape1 = 0.5,
  alpha_shape2 = 1,
  sigma_beta = 1
)
}
\arguments{
\item{num_tree}{Number of trees to include in ensemble.}

\item{k}{Numerator in scale for half-Cauchy prior in terminal node prior standard deviation.}

\item{base}{Parameter penalizing new nodes in the branching process prior.}

\item{power}{Parameter penalizing tree depth in the branching process prior.}

\item{mu_mu}{Mean parameter in terminal node prior.}

\item{sigma_mu}{Standard deviation parameter in terminal node prior.}

\item{s}{Vector of splitting probabilities for predictors.}

\item{alpha}{Positive constant controlling the sparsity level.}

\item{alpha_scale}{Scale of the prior for \code{alpha}; if not provided, defaults to the number of predictors.}

\item{alpha_shape1}{Shape parameter for prior on \code{alpha}; if not provided, defaults to 0.5.}

\item{alpha_shape2}{Shape parameter for prior on \code{alpha}; if not provided, defaults to 1.0.}

\item{sigma_beta}{Proposal distribution variance for fixed effect coefficients.}
}
\value{
Returns a list containing the function arguments.
}
\description{
Create a list of hyperparameter values
}
