% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clbart.R
\name{Opts}
\alias{Opts}
\title{MCMC options for clbart}
\usage{
Opts(
  num_burn = 2500,
  num_thin = 1,
  num_save = 2500,
  num_chains = 1,
  update_sigma_mu = TRUE,
  update_mu_mu = FALSE,
  update_s = FALSE,
  update_alpha = FALSE,
  update_sigma_beta = TRUE,
  store_lambda = FALSE,
  seed = 1996
)
}
\arguments{
\item{num_burn}{Number of burn-in iterations for the chain.}

\item{num_thin}{Post burn-in thinning interval for the chain.}

\item{num_save}{Number of posterior samples to save; in total, \code{num_burn + num_save * num_thin} iterations are run.}

\item{num_chains}{Number of chains to run; values greater than 1 only work with mc_clbart.}

\item{update_sigma_mu}{If \code{TRUE}, \code{sigma_mu} is updated, with a half-Cauchy prior.}

\item{update_mu_mu}{If \code{TRUE}, \code{mu_mu} is updated, with a noninformative conjugate normal prior (typically not recommended).}

\item{update_s}{If \code{TRUE}, \code{s} is updated using the Dirichlet prior \eqn{s \sim D(\alpha / P, \ldots, \alpha / P)} where \eqn{P} is the number of covariates used in the BART component.}

\item{update_alpha}{If \code{TRUE}, \code{alpha} is updated using a scaled beta prime prior.}

\item{update_sigma_beta}{If \code{TRUE}, \code{sigma_beta} is updated to encourage optimal M-H acceptance rates for the vector of fixed coefficients.}

\item{store_lambda}{If \code{TRUE}, values of the BART predictor for each individual are stored. In practice, this can be heavy on system memory.}

\item{seed}{Random seed for reproducibility.}
}
\value{
Returns a list containing the function arguments
}
\description{
MCMC options for clbart
}
