% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd.R
\name{pd}
\alias{pd}
\title{Partial dependence for clbart}
\usage{
pd(x, f_hat, vars, K, f = NULL)
}
\arguments{
\item{x}{Data frame of covariates used in the model-fitting process.}

\item{f_hat}{Function that returns a posterior distribution of predictions given covariates \code{x}.}

\item{vars}{Character vector of covariate names to vary in partial dependence calculations.}

\item{K}{Numeric vector of unique values to evaluate each covariate in \code{vars} at. By default this value will be decreased to the number of unique values if necessary.}

\item{f}{(Optional) function that returns the true prediction for a set of covariates. Useful when conducting simulation studies.}
}
\value{
A data frame of posterior mean estimates and 95\% credible intervals.
}
\description{
Compute partial dependence statistics
}
\examples{
# fit is an object returned by clbart()
# w is a matrix of covariates used in the BART component during model-fitting
# f_hat <- function (x) get_forest_posterior_predictions(fit$forest, x)
# pdw1 <- pd(w, f_hat, vars = 'W1', K = 10)
}
