% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clbart-predict-method.R
\name{predict.clbart}
\alias{predict.clbart}
\title{Predict Method for CL-BART Model Fits}
\usage{
\method{predict}{clbart}(object, newdata, type = "bart", posterior = FALSE)
}
\arguments{
\item{object}{a fitted object of class \code{clbart}.}

\item{newdata}{optionally, a list of data frames to be used for prediction.
When \code{type = "con"} is requested, this list must include a data frame of
covariates \code{x}. When \code{type = "bart"} is requested, this list must
include a data frame of exposure moderating covariates \code{w}.}

\item{type}{the type of prediction requested. The default (\code{"bart"})
produces predictions from the BART component of the CL-BART model fit on
the log odds-ratio scale, and \code{type = "con"} returns estimates of the
parametric component of the CL-BART model fit on the log odds scale.}

\item{posterior}{logical indicating whether or not the entire posterior
distribution for the requested predictions be returned. if \code{TRUE}
(default), a matrix is returned where each column represents the predictions
for a single posterior sample. If \code{FALSE}, a vector of posterior means
is returned.}
}
\value{
a vector (or matrix if \code{posterior = TRUE}) of predictions from
a fitted \code{clbart} model object.
}
\description{
Obtain predictions from a fitted \code{clbart} model.
}
\examples{
\dontrun{
# Assuming `model` is a fitted model of class `clbart`, and `w` is a data
# frame of modifiers of interest.
# To obtain the posterior distribution of predictions of exposure effects:
predict(model, list(w = w), type = 'bart', posterior = TRUE)
}
}
