% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clbart-summary-method.R
\name{summary.clbart}
\alias{summary.clbart}
\title{Summary Method for CL-BART Fits}
\usage{
\method{summary}{clbart}(object, ...)
}
\arguments{
\item{object}{a fitted object of class \code{clbart}.}

\item{...}{additional arguments affecting the summary produced.
\itemize{
\item \code{digits}: integer, specifying the number of decimal places (default is 2).
}}
}
\value{
A list of basic summaries for different components of a fitted
\code{clbart} model. The summary includes:
\itemize{
\item posterior means and credible intervals for traditional regression
coefficients (on the log odds-ratio scale).
\item posterior mean and credible interval for the overal average exposure effect
(on the log odds-ratio scale).
\item variable importance measures including posterior mean split proportions,
posterior mean splitting probabilities, and posterior inclusion probabilities
for each exposure effect moderator.
\item additional information (e.g., WAIC).
}
}
\description{
Produces basic summary statistics obtained from a fitted \code{clbart} model.
}
\examples{
\dontrun{
# Assuming `model` is a fitted model of class `clbart`
summary(model, digits = 3)
}
}
