% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clbart-package.R
\docType{package}
\name{clbart-package}
\alias{clbart-package}
\title{clbart: Conditional Logistic Regression with Bayesian Additive Regression Trees}
\description{
Estimate heterogeneous treatment/exposures effects in case-crossover studies using conditional logistic regression and Bayesian additive regression trees.
}
\author{
\strong{Maintainer}: Jacob Englert \email{jacob.englert@emory.edu}

}
\keyword{internal}
