% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clbart.R
\name{clbart}
\alias{clbart}
\title{Estimate Heterogeneous Effects with Conditional Logistic Regression}
\usage{
clbart(
  w,
  x = NULL,
  y,
  z,
  stratum,
  num_trees = 5,
  seed = 2187,
  iter = 10000,
  thin = 1,
  warmup = floor(iter/2),
  sigma2_beta = 1,
  sigma2_beta_update_freq = 100,
  beta_acc_prob = 0.35,
  n_min = 5,
  moves = c("grow", "prune", "change"),
  move_probs = c(0.3, 0.3, 0.4),
  alpha_rho = 0.95,
  beta_rho = 2,
  alpha_sigma = 1e-04,
  beta_sigma = 1e-04
)
}
\arguments{
\item{w}{A data frame of potential effect moderators.}

\item{x}{A data frame of effect confounders.}

\item{y}{A numeric vector of 1's and 0's corresponding to the outcome}

\item{z}{A numeric vector corresponding to the primary treatment or exposure
of interest. May be continuous or binary.}

\item{stratum}{A vector indicating the strata each observation belongs to.}

\item{num_trees}{integer number of trees to include in the BART portion of
the model}

\item{seed}{A random seed to initiate the algorithm with.}

\item{iter}{The total number of MCMC iterations to run.}

\item{thin}{Thinning rate for posterior samples.}

\item{warmup}{Number of MCMC iterations to run before storing posterior
samples.}

\item{sigma2_beta}{Initial value of the variance parameter used in the joint
multivariate normal proposal distribution for the confounding parameters.
This is updated during the warm-up period every \code{sigma2_beta_update_freq}
iterations.}

\item{sigma2_beta_update_freq}{Frequency with which to update \code{sigma2_beta}
during the warm-up period.}

\item{beta_acc_prob}{Target acceptance probability for the joint
Metropolis-Hastings proposal for the confounding parameters. \code{sigma2_beta}
will be updated throughout the MCMC with the goal of achieving this rate.}

\item{n_min}{Minimum number of cases for proposed terminal nodes. Moves that
would result in fewer than \code{n_min} cases in a leaf node are not considered.}

\item{moves}{Possible moves for the reversible-jump algorithm (DO NOT TOUCH).}

\item{move_probs}{Prior probabilities corresponding to the moves in \code{moves}.}

\item{alpha_rho, beta_rho}{Parameters for the tree splitting prior as defined
in Chipman et al. (2010).}

\item{alpha_sigma, beta_sigma}{Shape and scale parameters for the inverse
gamma prior on the terminal node variance.}
}
\value{
\code{clbart} returns a list of elements of the posterior distribution
generated during the fitting process. The number of posterior samples is
K = \code{(iter - warmup) / thin}.
\item{beta}{A K x \code{ncol(x)} matrix.}
\item{sigma2_beta}{A length K vector (typically one value, since sigma2_beta
is only updated during the warmup period).}
\item{beta_acc_rate}{A length K vector of the cumulative M-H acceptance rate
for the entire \code{beta} vector.}
\item{forests}{A length K list of lists. Each internal list represents an
individual tree within the ensemble, represented by a list of nodes within
within that tree for the given iteration. Each node is represented as a list
containing: \code{leaf} (indicates whether the node is a leaf node), \code{split_rule}
(indicating the variable and cut-point defining the node), and \code{mu}
(the prediction for the node).}
\item{tree_acc_rate}{A length K vector of the proportion of tree proposals
accepted out of \code{num_trees} within each iteration.}
\item{sigma2_mu}{A length K vector of the variance parameter placed on the
Normal leaf node prior.}
\item{split_props}{A K x \code{ncol(w)} matrix containing the distribution of
all splits in the ensemble across the variables in \code{w} at each iteration.}
\item{tree_props}{A K x \code{ncol(w)} matrix containing the proportion of trees
in the ensemble splitting on each variable in \code{w} at each iteration.}
\item{avg_tree_depth}{A length K vector of the average depth of all trees in
the ensemble at each iteration.}
\item{avg_num_nodes}{A length K vector of the average number of nodes in all
trees in the ensemble at each iteration.}
\item{avg_num_leaves}{A length K vector of the average number of leaves in
all trees in the ensemble at each iteration.}
\item{time}{A length K vector of the time taken to complete each iteration.}
\item{logLik}{A length K vector of the log-likelihood computed at each iteration.}
\item{pWAIC, lppd, WAIC}{The WAIC penalty, log posterior predictive density,
and WAIC (Widely Applicable Information Criteria) computed as:
\code{WAIC = -2*lppd +2*pWAIC}}
\item{...}{(the list also includes the parameters provided when fitting
the function.)}
}
\description{
Estimate heterogeneous effects of a single exposure/treatment covariate with
conditional logistic regression.
}
\examples{
'INSERT EXAMPLE HERE'

}
\references{
Chipman, H. A., George, E. I., and McCulloch, R. E. (2010). BART: Bayesian additive
regression trees. The Annals of Applied Statistics, 4(1).
}
