% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clbart.R
\name{clbart}
\alias{clbart}
\title{Fits the clbart model}
\usage{
clbart(w, x, y, z, strata, hypers, opts)
}
\arguments{
\item{w}{A matrix of covariates to include in the BART component.}

\item{x}{A matrix of covariates to include in the linear predictor component.}

\item{y}{A vector of outcomes (binary).}

\item{z}{A vector of primary exposures. This may be continuous or binary, but
must vary within each strata or else errors will occur.}

\item{strata}{An integer vector indicating the strata which each observation
belongs to. Prior to calling \code{clbart}, all data components must be
ordered by strata, and strata should be in increasing order.}

\item{hypers}{A list of hyperparameter values obtained from the \code{Hypers} function.}

\item{opts}{A list of MCMC options obtained from the \code{Opts} function.}
}
\value{
Returns a list with the following components:
}
\description{
Runs the CL-BART model for data which have been processed using the standard
(time-stratified) case-crossover approach.
}
\examples{
# TBD
}
